[graphql] Input.isDynamicCollect & Solid.isDynamicMapped
Summary:
Add a flag to the input dependency type to signify that it is a collecting input, and one to solid invocations to indicate that they are downstream of a dynamic output.

These bits of information need to get added to the snapshot as well, so ids change.

Test Plan: added test

Reviewers: bengotow, prha, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D6940