[execute_in_process improvements 1/n] make scoped resources work with type annotation
Summary:
The resource object that we expose to users is a dynamically-attributed namedtuple. This actually results in a type annotation error. Additionally, because we dynamically generate the type, we don't have a more specific type to use for annotation than Any. This diff adds a "tag" type that we can use to annotate locations where we use this resources namedtuple instead of just using Any.

This is directly useful for using build_resources with execute_in_process: https://dagster.phacility.com/D6822

Test Plan: Unit tests. Ensure behavior doesn't change anywhere.

Reviewers: sandyryza, alangenfeld, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D7060