prevent start/stop of sensors from messing up tick timestamps
Summary:
We were writing tick timestamps when we were turning sensors on/off.  That already doesn't
make sense, but we were also writing it in a different format (i.e. not pendulum.now('UTC')).

This diff makes it so that tick timestamps are only written in one place and on/off operations
just affect status.

This was the root cause for the weird behavior listed here: https://dagster.slack.com/archives/CCCR6P2UR/p1615994177024300

Test Plan: added test for sensor on/off behavior

Reviewers: dgibson, sashank, alangenfeld

Reviewed By: alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D7071