Add check for RabbitMQ in Celery deployment
Summary:
In the kubernetes event logs, there were errors related to the liveness probe which may have been causing the celery user deployment tests to flake.

> looking at the event log errors in the flaking tests: looks like the celery workers are failing a health check after being scheduled onto a node
> i.e. for https://buildkite.com/dagster/dagster-integration-tests/builds/1844#0707661d-c836-4663-890c-1e39a199bfbf
> we can look at the created namespace’s event log to see what’s happening
> https://s3.amazonaws.com/buildkiteartifacts.com/a43ea0e6-ae44-489a-b2a4-ff6149cb3238/[…]773ad82a2192b5422cf2e6b50b631bf63c24208f7f2c97b4e8b4444b2c79

After removing the liveness probe, the flakiness has seemingly disappeared...

Turns out, the liveness probe was failing since RabbitMQ is sometimes not ready when the probe starts running. To resolve, we add an explicit dependency on RabbitMQ in the Celery deployment in the form of an init container.

Test Plan: integration

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D7076