[dagit] Fix cache reading in Schedules/Sensors section
Summary:
Now that the left nav allows multiple repositories to be selected, the Schedules/Sensors section does not update status correctly. This is because the queries are manually performed in an array of `client.query` calls, and do not subscribe to the cache.

Repair this by switching to `watchQuery` subscriptions, which will observe the cache. When data arrives, update interal state of `JobsList`.

Test Plan: Run dagit, switch schedules and sensors on and off. Verify that the status dot is shown or hidden accordingly.

Reviewers: dgibson, bengotow, prha

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D7220