[dagit] RFC: Simplify GraphQL Playground usage
Summary:
We currently require graphql-playground-react in `dagit/core`, but only to grab its static assets from node_modules and jam them into a vendor directory to be referenced in `playground.py`.

Instead, check in the static assets and just `cp` them to the target `vendor` directory during the `build-for-python` step.

Open to discussion on this.

Test Plan: Run dagit, navigate to `/graphql` locally. Verify that the GraphQL playground loads and behaves correctly.

Reviewers: bengotow, prha, sashank, max

Reviewed By: bengotow, max

Differential Revision: https://dagster.phacility.com/D7284