Switch partition run matrix to simpler “square color = status in last run”
Summary:
This addresses https://github.com/dagster-io/dagster/issues/3792 - instead of a fancy "green if this step has succeeded in any run of this partition", we colorize the partition run squares to show the result of the most recent step execution.

In the advanced "show previous states" mode, we split the box and show the previous status if there was one that was different from the final status. (Eg: if it failed and then succeeded, or succeeded and then failed, etc.) I'm not sure this is perfect but it's now an advanced option anyway.

The only slightly odd part of this is "SKIPPED". I think this is a valid execution result (?) so a step that succeeded in Run 1 and then skipped in Run 2 will show as yellow / skipped unless you enable the previous state mode and see green/yellow.

Test Plan: Give partition matrix a spin!

Reviewers: sandyryza

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D7377