in check functions, remove raise_with_traceback and convert to f-strings
Summary:
The primary motivation for this change is to remove an un-useful frame from the stack traces that users see when they hit a check error.

That change uncovered a mypy error, which I fixed by replacing Set with AbstractSet in a bunch of different places.

While on the theme of python 3 stuff, I made some format->f-string conversions in the check module.  I ran out of steam before completing all of them, but I still think this is an improvement.

Test Plan: bk

Reviewers: alangenfeld, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D7499