Better error handling when the daemon hits an error while trying to check a heartbeat
Summary:
The daemon periodically checks heartbeat status to identify when a daemon is hanging and needs to be restarted, but if you get unlucky and happen to hit a transient error in this heartbeat check, the daemon will immediately assume that it has failed and crash.

Instead, allow a period of heartbeat check failures before deciding to restart the daemon. This way, we still identify when the daemon is hanging or not reaching the DB for a longperiod of time, but we are more resilient to transient failures.

Test Plan: BK

Reviewers: johann, prha

Reviewed By: johann

Subscribers: max

Differential Revision: https://dagster.phacility.com/D7535