[dagit] Make log timestamps permalinks
Summary:
Resolves #4068.

Allow linking directly to a log entry on a Run, using the `focusedTime` parameter. It looks like this probably used to work, and might have gotten lost in the shuffle when we refactored filtering behavior.

The timestamp of a log row is now clickable, and the row will now be highlighted and scrolled into view.

Test Plan: View a run. Click on timestamp in log row, verify highlighting. Reload the page, verify that the query parameter was set in the URL, and that the log is scrolled into view immediately.

Reviewers: bengotow, prha, sandyryza

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D7540