Heartbeat right away in the daemon on startup
Summary: Adding 30 seconds before any heartbeats are logged in D7494 feels like a mistake - particularly now that errors are not considered to be equated with 'not running'. Better to get immediate feedback that the daemon is running, then 30 seconds later get any errors that came in during the first iteration.

Test Plan: Integration

Reviewers: johann, prha

Reviewed By: johann

Differential Revision: https://dagster.phacility.com/D7600