[RFC] Construct partitions using serializable parameters
Summary:
Depends on D7642.

Rather than construct the partitions using a user specified closure, we
want to capture all the relevant serializable information that can be used to
potentially generate the closure at run time when partition names are being retrieved.

We envision three classes of parameters to allow for this:
- static
- time based (e.g. a date partition)
- dynamic (the current original behavior)

Test Plan: N/A

Reviewers: dgibson, sidkmenon, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D7621