Re-rache repository locations on WorkspaceProcessContext everytime a request is made
Summary:
Previously, the repository locations from the workspace were cached on the `WorkspaceProcessContext` when initialized. This is because the process context assumes that the workspace only changes when `reload_workspace` is called on the `ProcessContext`. With the new async workspaces, this is not true anymore.

This diff changes the process context to not cache the locations anymore and directly access the underlying workspace instead

Test Plan: Existing tests

Reviewers: dgibson

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D7682