[dagit] Make Non-Subscription GraphQL ops use HTTP
Summary:
As part of an ask by a user (and generally for best practice), we are moving non-subscription GraphQL operations to use HTTP instead of a websocket connection.

The removal of the GEventExecutor is covered in [[ https://github.com/dagster-io/dagster/issues/4143 | GH issue #4143 ]]

Test Plan:
bk

Also checked that the app behaves similarly to dagit built off of using only websockets (i.e. static JS built from master, dev JS built from branch).

When the webserver is killed, dagit does not immediately crash or anything. However, a few new behaviors are seen:

1. As in the picture below, we can see that a red error bar is shown in the top right-hand corner, and the 'Runs' page renders a bit differently when the backing GraphQL server is offline:

{F776754}

These behaviors are not seen on the production dagit, which looks like this - the only indication that the server is down,besides the gray circle in the corner is console print statements

{F776764}

This is behavior that is due to the difference between websockets & http - the websocket response does not have an 'error' field on the JS side so the components render differently. Not sure if throwing away the error in this case (to conform to the old behavior) makes sense though -- feels like errors are important to pay attention to!

2. (related) Some code paths that were not utilized before are now exercised. As detailed in https://github.com/apollographql/apollo-link/issues/547, the `onError` link does not run properly with a WebSocketLink - therefore, the logic set up [[ https://sourcegraph.com/github.com/dagster-io/dagster/-/blob/js_modules/dagit/packages/core/src/app/AppError.tsx#L38-43 | here ]] that creates the red banner in the corner in this version of Dagit upon network failure does not render at all / is dead code in the websocket-only Dagit.

3. Checked that moving HTTP to the SyncExecutor still works fine - triggered a run using the `toys/many_events` pipeline, kept the Gantt chart open for that in one tab (which requires a subscription), and then clicked around as many panes in dagit as I could. Everything was still snappy and loaded fine.

Reviewers: dish, bengotow, alangenfeld, sashank, dgibson

Reviewed By: dish, alangenfeld

Differential Revision: https://dagster.phacility.com/D7688