[docs] util script that removes versioned content
Summary:
this could be used in the release pipeline where when it's a pre-release we remove the rc versioned doc at the end.

this diff chose to implement the removal rather than skipping generating docs for rc, bc i think we do want to test every step in a pre-release which includes updating the docs version.

the cmd is:
`python remove_version.py --version <version>`

Test Plan:
`python remove_version.py --version <version>`
{F779739}

Reviewers: max, sashank, alangenfeld

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D7784