Load new image into existing kind cluster
Summary: When iterating on kind tests locally, it's much faster to reuse the same kind cluster. This change supports that workflow by loading in new images when the integration image has changed. If it hasn't changed, kind recognizes by the hash and skips loading it itself.

Test Plan: local integration

Reviewers: dgibson, jordansanders

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D7903