[dagit] Fix config nubs
Summary:
Resolves #4138

A recent change made it so that the DAG viewer rendered config nubs on every solid, instead of just solids with configs. Repair this by skipping the nub for any solids with `Any` config.

Test Plan: View a bunch of different pipelines, verify that the config nub only appears on relevant solids.

Reviewers: bengotow, prha

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D7912