[RFC] Context managers for building input/output contexts
Summary:
Right now, we don't have a compelling testing story for IO managers. In order to test the functionality of an IO manager class, you need to manually construct InputContext and OutputContext, but there is no simple way for a user to provide resources to these. In addition, having a user directly create contexts makes it harder to support changing them around. This diff adds `build_input_context` and `build_output_context` to try to address these issues.

One possibly contentious part here is the context manager stuff that I have done to InputContext/OutputContext. I think, as long as we want either the classes themselves or sublcasses to be context managers, we need to move out of the namedtuple syntax.

Test Plan:
Unit, and migrating over existing tests to see if they improve anything

API Reference:
{F820922}

{F820921}

Reviewers: sandyryza, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D7926